refactor: move `fetch_github_release_config`
